[Bevy] Load BevyASM and dependency as an asset
The link between BevyASM and its frame_source is not as strong
as I would ideally like it to be in the example script.
i.e. While loading the asm does also load the texture atlas the
atlas also needs to be manually loaded in order to add its handle
to the SpriteSheetBundle